+++ src/main/kotlin/com/asap/asapbackend/batch/classroom/ClassroomScheduler.kt
@@ -1,17 +1,16 @@
-package com.asap.asapbackend.batch.school.application
+package com.asap.asapbackend.batch.classroom
 
-import jakarta.annotation.PostConstruct
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-class ShoolScheduler(
-        private val schoolService: SchoolService
+class ClassroomScheduler(
+        private val classroomService: ClassroomService
 ) {
     @Scheduled(cron = "0 0 4 1 3 ?") // 매년 3월 1일 04:00:00에 실행
-    @GetMapping("/addSchool")
-    fun addSchool() {
-        schoolService.addSchool()
+    @GetMapping("/addClassroom")
+    fun addClassroom(){
+        classroomService.addClassroom()
     }
 }
