+++ src/main/kotlin/com/asap/asapbackend/batch/classroom/ClassroomService.kt
@@ -1,13 +1,12 @@
-package com.asap.asapbackend.batch.classroom.application
+package com.asap.asapbackend.batch.classroom
 
 import com.asap.asapbackend.domain.classroom.domain.service.ClassroomAppender
-import com.asap.asapbackend.batch.classroom.service.ClassroomOpenApiClient
 import org.springframework.stereotype.Service
 
 @Service
 class ClassroomService(
-        private val classroomReader: ClassroomOpenApiClient,
-        private val classroomAppender: ClassroomAppender
+    private val classroomReader: ClassroomOpenApiClient,
+    private val classroomAppender: ClassroomAppender
 ) {
     fun addClassroom() {
         val classroom = classroomReader.getClassroom()
