+++ src/main/kotlin/com/asap/asapbackend/batch/classroom/dto/ClassInfo.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.classroom.model
+package com.asap.asapbackend.batch.classroom.dto
 
 data class ClassInfo(
         val classInfo: List<ClassRoomInfo>
