+++ src/main/kotlin/com/asap/asapbackend/batch/classroom/dto/ClassroomResponse.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.classroom.model
+package com.asap.asapbackend.batch.classroom.dto
 
 import com.asap.asapbackend.domain.school.domain.model.School
