+++ src/main/kotlin/com/asap/asapbackend/batch/school/ShoolScheduler.kt
@@ -1,16 +1,16 @@
-package com.asap.asapbackend.batch.classroom.application
+package com.asap.asapbackend.batch.school
 
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-class ClassroomScheduler(
-        private val classroomService: ClassroomService
+class ShoolScheduler(
+        private val schoolService: SchoolService
 ) {
     @Scheduled(cron = "0 0 4 1 3 ?") // 매년 3월 1일 04:00:00에 실행
-    @GetMapping("/addClassroom")
-    fun addClassroom(){
-        classroomService.addClassroom()
+    @GetMapping("/addSchool")
+    fun addSchool() {
+        schoolService.addSchool()
     }
 }
