+++ src/main/kotlin/com/asap/asapbackend/batch/school/dto/NeisSchoolInfoJS.kt
@@ -1,4 +1,4 @@
-package com.asap.asapbackend.batch.school.model
+package com.asap.asapbackend.batch.school.dto
 
 data class NeisSchoolInfoJS(
         val neisSchoolInfoJS: List<SchoolInfo>
